Moved wrappers of system classes to Core
Changed namespaces. Fixed up existing code.
No code changes.
