+++ src/Core/SystemAbstractions/IO/DirectoryWrapper.cs
@@ -18,19 +18,24 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
 
-namespace SonarLint.VisualStudio.Integration
+namespace SonarLint.VisualStudio.Core.SystemAbstractions
 {
-    public interface ITimer : IDisposable
+    [ExcludeFromCodeCoverage] // Wrapper around System
+    public class DirectoryWrapper : IDirectory
     {
-        event EventHandler<TimerEventArgs> Elapsed;
+        public void Create(string path) => Directory.CreateDirectory(path);
 
-        bool AutoReset { get; set; }
-        double Interval { get; set; }
+        public void Delete(string path) => Directory.Delete(path);
 
-        void Start();
-        void Stop();
-        void Close();
+        public void Delete(string path, bool recursive) => Directory.Delete(path, recursive);
+
+        public bool Exists(string path) => Directory.Exists(path);
+
+        public string[] GetDirectories(string path, string searchPattern) => Directory.GetDirectories(path, searchPattern);
+
+        public void Move(string sourceDirName, string destDirName) => Directory.Move(sourceDirName, destDirName);
     }
 }
