+++ src/Core/SystemAbstractions/IO/FileWrapper.cs
@@ -18,10 +18,24 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarLint.VisualStudio.Integration.Helpers
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
+
+namespace SonarLint.VisualStudio.Core.SystemAbstractions
 {
-    public interface IFileSystemWatcherFactory
+    [ExcludeFromCodeCoverage] // Wrapper around System.IO.File
+    public class FileWrapper : IFile
     {
-        IFileSystemWatcher Create();
+        public void Delete(string path) => File.Delete(path);
+
+        public bool Exists(string path) => File.Exists(path);
+
+        public TextReader OpenText(string path) => File.OpenText(path);
+
+        public TextWriter CreateText(string path) => File.CreateText(path);
+
+        public string ReadAllText(string path) => File.ReadAllText(path);
+
+        public void WriteAllText(string path, string contents) => File.WriteAllText(path, contents);
     }
 }
