+++ src/Core/SystemAbstractions/IO/IFile.cs
@@ -18,24 +18,17 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
 
-namespace SonarLint.VisualStudio.Integration.Helpers
+namespace SonarLint.VisualStudio.Core.SystemAbstractions
 {
-    [ExcludeFromCodeCoverage] // Wrapper around System.IO.File
-    public class FileWrapper : IFile
+    public interface IFile
     {
-        public void Delete(string path) => File.Delete(path);
-
-        public bool Exists(string path) => File.Exists(path);
-
-        public TextReader OpenText(string path) => File.OpenText(path);
-
-        public TextWriter CreateText(string path) => File.CreateText(path);
-
-        public string ReadAllText(string path) => File.ReadAllText(path);
-
-        public void WriteAllText(string path, string contents) => File.WriteAllText(path, contents);
+        bool Exists(string path);
+        void Delete(string path);
+        TextReader OpenText(string path);
+        TextWriter CreateText(string path);
+        string ReadAllText(string path);
+        void WriteAllText(string path, string contents);
     }
 }
