+++ src/Core/SystemAbstractions/Timer/ITimerFactory.cs
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-namespace SonarLint.VisualStudio.Integration
+namespace SonarLint.VisualStudio.Core.SystemAbstractions
 {
     public interface ITimerFactory
     {
