+++ src/Integration.UnitTests/Notifications/SonarQubeNotificationServiceTests.cs
@@ -29,6 +29,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SonarQube.Client.Models;
 using SonarQube.Client;
+using SonarLint.VisualStudio.Core.SystemAbstractions;
 
 namespace SonarLint.VisualStudio.Integration.Notifications.UnitTests
 {
