+++ src/Integration.Vsix/SonarLintDaemon/LegacyInstallationCleanup.cs
@@ -20,7 +20,7 @@
 
 using System;
 using System.IO;
-using SonarLint.VisualStudio.Integration.Helpers;
+using SonarLint.VisualStudio.Core.SystemAbstractions;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
