+++ src/Integration/Notifications/SonarQubeNotificationService.cs
@@ -22,6 +22,7 @@ using System;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
+using SonarLint.VisualStudio.Core.SystemAbstractions;
 using SonarLint.VisualStudio.Integration.Resources;
 using SonarQube.Client;
 using CancellationTokenSource = System.Threading.CancellationTokenSource;
