+++ src/Integration/Persistence/FileBindingSerializer.cs
@@ -23,6 +23,7 @@ using System.Diagnostics;
 using System.IO;
 using Microsoft.Alm.Authentication;
 using Newtonsoft.Json;
+using SonarLint.VisualStudio.Core.SystemAbstractions;
 using SonarLint.VisualStudio.Integration.Helpers;
 using SonarLint.VisualStudio.Integration.Resources;
 using SonarQube.Client.Helpers;
